pyproject.toml
[tool.poetry]
name = "educationservice"
version = "0.1.0"
description = ""
authors = ["AndreyKosov <andrey.kosov.05@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.6"}
aiohttp = "^3.11.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
